+++ rotkehlchen/chain/ethereum/adex/adex.py
@@ -25,12 +25,13 @@ from typing_extensions import Literal
 from web3 import Web3
 
 from rotkehlchen.accounting.structures import Balance
-from rotkehlchen.assets.asset import EthereumToken
 from rotkehlchen.chain.ethereum.graph import GRAPH_QUERY_LIMIT, Graph, format_query_indentation
 from rotkehlchen.chain.ethereum.utils import generate_address_via_create2
-from rotkehlchen.constants.assets import A_ADX
+from rotkehlchen.constants import YEAR_IN_SECONDS, ZERO
+from rotkehlchen.constants.assets import A_ADX, A_DAI, A_USD
 from rotkehlchen.errors import RemoteError
 from rotkehlchen.fval import FVal
+from rotkehlchen.history import PriceHistorian
 from rotkehlchen.inquirer import Inquirer
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.premium.premium import Premium
@@ -39,29 +40,32 @@ from rotkehlchen.user_messages import MessagesAggregator
 from rotkehlchen.utils.interfaces import EthereumModule
 from rotkehlchen.utils.serialization import rlk_jsonloads_list
 
-from .graph import BONDS_QUERY, UNBOND_REQUESTS_QUERY, UNBONDS_QUERY
+from .graph import BONDS_QUERY, CHANNEL_WITHDRAWS_QUERY, UNBOND_REQUESTS_QUERY, UNBONDS_QUERY
 from .typing import (
     ADXStakingBalance,
+    ADXStakingDetail,
     ADXStakingEvents,
     ADXStakingHistory,
-    ADXStakingStat,
     Bond,
+    ChannelWithdraw,
     DeserializationMethod,
     EventCoreData,
     EventType,
     TomPoolIncentive,
     Unbond,
     UnbondRequest,
+    UnclaimedReward,
 )
 from .utils import (
     ADEX_EVENTS_PREFIX,
     ADX_AMOUNT_MANTISSA,
     CREATE2_SALT,
+    DAI_AMOUNT_MANTISSA,
+    EVENT_TYPE_ORDER_IN_ASC,
     IDENTITY_FACTORY_ADDR,
     IDENTITY_PROXY_INIT_CODE,
     PERIOD_END_AT_FORMAT,
     POOL_ID_POOL_NAME,
-    SECONDS_PER_YEAR,
     STAKING_ADDR,
     TOM_POOL_FEE_REWARDS_API_URL,
     TOM_POOL_ID,
@@ -112,7 +116,10 @@ class Adex(EthereumModule):
             bonds: List[Bond],
             unbonds: List[Unbond],
             unbond_requests: List[UnbondRequest],
+            channel_withdraws: List[ChannelWithdraw],
+            unclaimed_rewards: Dict[ChecksumAddress, UnclaimedReward],
             adx_usd_price: Price,
+            dai_usd_price: Price,
     ) -> Dict[ChecksumAddress, List[ADXStakingBalance]]:
         """Given a list of bonds, unbonds and unbond requests returns per
         address the staked amounts per pool.
@@ -125,6 +132,10 @@ class Adex(EthereumModule):
         address_unbond_requests_set = {
             (unbond_request.address, unbond_request.bond_id) for unbond_request in unbond_requests
         }
+        address_channel_withdraws = defaultdict(list)
+        for channel_withdraw in channel_withdraws:
+            address_channel_withdraws[channel_withdraw.address].append(channel_withdraw)
+
         # Get bonds whose `bond_id` is not in unbonds or unbond_requests
         for bond in bonds:
             if (
@@ -139,21 +150,45 @@ class Adex(EthereumModule):
         for address, bonds in address_bonds.items():
             pool_ids = {bond.pool_id for bond in bonds}
             for pool_id in pool_ids:
-                amount = FVal(sum(bond.amount for bond in bonds if bond.pool_id == pool_id))
                 pool_name = POOL_ID_POOL_NAME.get(pool_id, None)
                 if pool_name is None:
                     log.error(
                         f'Error getting name for AdEx pool: {pool_id}. '
-                        'Please, update the map of pools and names.',
+                        f'Please, update the map of pools and names.',
                     )
+                # Add the ADX amount staked
+                adx_staked_amount = FVal(
+                    sum(bond.value.amount for bond in bonds if bond.pool_id == pool_id),
+                )
+                # Add the ADX amount claimed
+                adx_claimed_amount = FVal(
+                    sum(event.value.amount for event in address_channel_withdraws.get(address, [])),  # noqa: E501
+                )
+                # Add ADX and DAI from the unclaimed feed
+                adx_unclaimed_amount = ZERO
+                dai_unclaimed_amount = ZERO
+                if address in unclaimed_rewards:
+                    adx_unclaimed_amount = (
+                        unclaimed_rewards[address].adx_amount - adx_claimed_amount
+                    )
+                    dai_unclaimed_amount = unclaimed_rewards[address].dai_amount
+
                 pool_balance = ADXStakingBalance(
                     pool_id=pool_id,
                     pool_name=pool_name,
-                    balance=Balance(
-                        amount=amount,
-                        usd_value=amount * adx_usd_price,
+                    adx_balance=Balance(
+                        amount=adx_staked_amount + adx_unclaimed_amount,
+                        usd_value=(adx_staked_amount + adx_unclaimed_amount) * adx_usd_price,
                     ),
-                    address=to_checksum_address(STAKING_ADDR),
+                    adx_unclaimed_balance=Balance(
+                        amount=adx_unclaimed_amount,
+                        usd_value=adx_unclaimed_amount * adx_usd_price,
+                    ),
+                    dai_balance=Balance(
+                        amount=dai_unclaimed_amount,
+                        usd_value=dai_unclaimed_amount * dai_usd_price,
+                    ),
+                    contract_address=to_checksum_address(STAKING_ADDR),
                 )
                 adex_balances[address].append(pool_balance)
 
@@ -171,8 +206,8 @@ class Adex(EthereumModule):
           type.
           - tom_pool_incentive (optional): Tom pool incentive data.
 
-        Return a map between an address and its <ADXStakingStat>, which contains
-        all the events that belong to the address, and the performance stats
+        Return a map between an address and its <ADXStakingDetail>, which contains
+        all the events that belong to the address, and the performance details
         per staking pool.
         """
         staking_history = {}
@@ -180,35 +215,55 @@ class Adex(EthereumModule):
         all_events: List[Union[Bond, Unbond, UnbondRequest]] = (
             staking_events.bonds +
             staking_events.unbonds +  # type: ignore # mypy bug concatenating lists
-            staking_events.unbond_requests  # type: ignore # mypy bug concatenating lists
+            staking_events.unbond_requests +  # type: ignore # mypy bug concatenating lists
+            staking_events.channel_withdraws  # type: ignore # mypy bug concatenating lists
         )
         # Map addresses with their events
         for event in all_events:
             address_staking_events[event.address].append(event)
-        # Sort staking events per address by timestamp (older first)
+        # Sort staking events per address by timestamp (older first) and event type
         for address in address_staking_events.keys():
-            address_staking_events[address].sort(key=lambda event: event.timestamp)
+            address_staking_events[address].sort(
+                key=lambda event: (event.timestamp, EVENT_TYPE_ORDER_IN_ASC[type(event)]),
+            )
 
         for address, adx_staking_balances in staking_balances.items():
-            adx_staking_stats = []
+            adx_staking_details = []
             for adx_staking_balance in adx_staking_balances:
-                # NB: currently it only returns stats for Tom pool
+                # NB: currently it only returns staking details for Tom pool
                 if adx_staking_balance.pool_id == TOM_POOL_ID and tom_pool_incentive is not None:
-                    pool_staking_stat = ADXStakingStat(
-                        address=adx_staking_balance.address,
+                    total_profit_amount = FVal(ZERO)
+                    total_profit_usd_value = FVal(ZERO)
+                    # Add claimed amounts and their historical usd value
+                    for event in address_staking_events[address]:
+                        if type(event) == ChannelWithdraw:
+                            total_profit_amount += event.value.amount
+                            total_profit_usd_value += event.value.usd_value
+
+                    # Add unclaimed amount and its current usd value
+                    total_profit_amount += adx_staking_balance.adx_unclaimed_balance.amount
+                    total_profit_usd_value += adx_staking_balance.adx_unclaimed_balance.usd_value
+
+                    pool_staking_detail = ADXStakingDetail(
+                        contract_address=adx_staking_balance.contract_address,
                         pool_id=adx_staking_balance.pool_id,
                         pool_name=adx_staking_balance.pool_name,
                         total_staked_amount=(
                             tom_pool_incentive.total_staked_amount / ADX_AMOUNT_MANTISSA
                         ),
                         apr=tom_pool_incentive.apr,
-                        balance=adx_staking_balance.balance,
+                        adx_balance=adx_staking_balance.adx_balance,
+                        dai_balance=adx_staking_balance.dai_balance,
+                        profit_loss=Balance(
+                            amount=total_profit_amount,
+                            usd_value=total_profit_usd_value,
+                        ),
                     )
-                    adx_staking_stats.append(pool_staking_stat)
+                    adx_staking_details.append(pool_staking_detail)
 
             staking_history[address] = ADXStakingHistory(
                 events=address_staking_events[address],
-                staking_stats=adx_staking_stats,
+                staking_details=adx_staking_details,
             )
         return staking_history
 
@@ -241,10 +296,37 @@ class Adex(EthereumModule):
             identity_address=event_core_data.identity_address,
             timestamp=event_core_data.timestamp,
             bond_id=bond_id,
-            amount=amount,
+            value=Balance(amount=amount),
             pool_id=pool_id,
             nonce=nonce,
             slashed_at=Timestamp(int(raw_event['slashedAtStart'])),
+        )
+
+    @staticmethod
+    def _deserialize_channel_withdraw(
+            raw_event: Dict[str, Any],
+            identity_address_map: Dict[ChecksumAddress, ChecksumAddress],
+    ) -> ChannelWithdraw:
+        """Deserialize a channel withdraw event. Only for Tom pool.
+
+        It may raise KeyError.
+
+        TODO: use bidict or already pass the inverse identity_address_map.
+        """
+        inverse_identity_address_map = {
+            address: identity for identity, address in identity_address_map.items()
+        }
+        address = to_checksum_address(raw_event['user'])
+        identity_address = inverse_identity_address_map[address]
+        amount = FVal(raw_event['amount']) / ADX_AMOUNT_MANTISSA
+        return ChannelWithdraw(
+            tx_hash=HexStr(raw_event['id'].split(':')[0]),
+            address=address,
+            identity_address=identity_address,
+            timestamp=Timestamp(raw_event['timestamp']),
+            value=Balance(amount=amount),
+            channel_id=HexStr(raw_event['channelId']),
+            pool_id=TOM_POOL_ID,
         )
 
     @staticmethod
@@ -284,6 +366,7 @@ class Adex(EthereumModule):
             identity_address=event_core_data.identity_address,
             timestamp=event_core_data.timestamp,
             bond_id=HexStr(raw_event['bondId']),
+            value=Balance(),
         )
 
     def _deserialize_unbond_request(
@@ -305,6 +388,7 @@ class Adex(EthereumModule):
             identity_address=event_core_data.identity_address,
             timestamp=event_core_data.timestamp,
             bond_id=HexStr(raw_event['bondId']),
+            value=Balance(),
             unlock_at=Timestamp(int(raw_event['willUnlock'])),
         )
 
@@ -326,7 +410,7 @@ class Adex(EthereumModule):
 
         if response.status_code != HTTPStatus.OK:
             msg = (
-                'AdEx fee rewards API query responded with error status code: '
+                f'AdEx fee rewards API query responded with error status code: '
                 f'{response.status_code} and text: {response.text}.'
             )
             log.error(msg)
@@ -378,7 +462,7 @@ class Adex(EthereumModule):
                 min_from_timestamp = min(min_from_timestamp, events_range[1])
 
         # Request new addresses' events
-        all_new_events: List[Union[Bond, Unbond, UnbondRequest]] = []
+        all_new_events: List[Union[Bond, Unbond, UnbondRequest, ChannelWithdraw]] = []
         if new_addresses:
             new_events = self._get_new_staking_events_graph(
                 addresses=addresses,
@@ -398,6 +482,7 @@ class Adex(EthereumModule):
 
         # Add new events in DB
         if all_new_events:
+            self._update_staking_events_data(events=all_new_events)
             self.database.add_adex_events(all_new_events)
 
         # Fetch all DB events within the time range
@@ -441,53 +526,75 @@ class Adex(EthereumModule):
     ) -> List[UnbondRequest]:
         ...
 
+    @overload  # noqa: F811
+    def _get_staking_events_graph(  # pylint: disable=no-self-use
+            self,
+            addresses: List[ChecksumEthAddress],
+            event_type: Literal[EventType.CHANNEL_WITHDRAW],
+            from_timestamp: Optional[Timestamp] = None,
+            to_timestamp: Optional[Timestamp] = None,
+    ) -> List[ChannelWithdraw]:
+        ...
+
     def _get_staking_events_graph(
             self,
             addresses: List[ChecksumEthAddress],
-            event_type: Literal[EventType.BOND, EventType.UNBOND, EventType.UNBOND_REQUEST],
+            event_type: Literal[
+                EventType.BOND,
+                EventType.UNBOND,
+                EventType.UNBOND_REQUEST,
+                EventType.CHANNEL_WITHDRAW,
+            ],
             from_timestamp: Optional[Timestamp] = None,
             to_timestamp: Optional[Timestamp] = None,
-    ) -> Union[List[Bond], List[Unbond], List[UnbondRequest]]:
+    ) -> Union[List[Bond], List[Unbond], List[UnbondRequest], List[ChannelWithdraw]]:
         """Get the addresses' events data querying the AdEx subgraph.
         """
-        identity_address_map = (
-            self._get_identity_address_map(addresses)
-        )
+        identity_address_map = self._get_identity_address_map(addresses)
+        user_identities = [str(identity).lower() for identity in identity_address_map.keys()]
         deserialization_method: DeserializationMethod
         querystr: str
-        schema: Literal['bonds', 'unbonds', 'unbondRequests']
+        schema: Literal['bonds', 'unbonds', 'unbondRequests', 'channelWithdraws']
         if event_type == EventType.BOND:
+            queried_addresses = user_identities
             deserialization_method = self._deserialize_bond
             querystr = format_query_indentation(BONDS_QUERY.format())
             schema = 'bonds'
             event_type_pretty = 'bond'
         elif event_type == EventType.UNBOND:
+            queried_addresses = user_identities
             deserialization_method = self._deserialize_unbond
             querystr = format_query_indentation(UNBONDS_QUERY.format())
             schema = 'unbonds'
             event_type_pretty = 'unbond'
         elif event_type == EventType.UNBOND_REQUEST:
+            queried_addresses = user_identities
             deserialization_method = self._deserialize_unbond_request
             querystr = format_query_indentation(UNBOND_REQUESTS_QUERY.format())
             schema = 'unbondRequests'
             event_type_pretty = 'unbond request'
+        elif event_type == EventType.CHANNEL_WITHDRAW:
+            queried_addresses = [address.lower() for address in addresses]
+            deserialization_method = self._deserialize_channel_withdraw
+            querystr = format_query_indentation(CHANNEL_WITHDRAWS_QUERY.format())
+            schema = 'channelWithdraws'
+            event_type_pretty = 'channel withdraws'
         else:
             raise AssertionError(f'Unexpected AdEx event type: {event_type}.')
 
-        user_identities = [str(identity).lower() for identity in identity_address_map.keys()]
         start_ts = from_timestamp or 0
         end_ts = to_timestamp or int(datetime.utcnow().timestamp())
         param_types = {
             '$limit': 'Int!',
             '$offset': 'Int!',
-            '$user_identities': '[Bytes!]',
+            '$queried_addresses': '[Bytes!]',
             '$start_ts': 'Int!',
             '$end_ts': 'Int!',
         }
         param_values = {
             'limit': GRAPH_QUERY_LIMIT,
             'offset': 0,
-            'user_identities': user_identities,
+            'queried_addresses': queried_addresses,
             'start_ts': start_ts,
             'end_ts': end_ts,
         }
@@ -538,11 +645,11 @@ class Adex(EthereumModule):
                 'offset': offset + GRAPH_QUERY_LIMIT,
             }
 
-        return events  # type: ignore # type is not List[Union[Bond, Unbond, UnbondRequest]]
+        return events  # type: ignore # the suggested type is not the right one
 
     @staticmethod
     def _get_addresses_staking_events_grouped_by_type(
-            events: List[Union[Bond, Unbond, UnbondRequest]],
+            events: List[Union[Bond, Unbond, UnbondRequest, ChannelWithdraw]],
             addresses: Set[ChecksumAddress],
     ) -> ADXStakingEvents:
         """Filter out events that don't belong to any of the addresses and
@@ -551,6 +658,7 @@ class Adex(EthereumModule):
         bonds = []
         unbonds = []
         unbond_requests = []
+        channel_withdraws = []
         for event in events:
             if event.address in addresses:
                 if isinstance(event, Bond):
@@ -559,6 +667,8 @@ class Adex(EthereumModule):
                     unbonds.append(event)
                 elif isinstance(event, UnbondRequest):
                     unbond_requests.append(event)
+                elif isinstance(event, ChannelWithdraw):
+                    channel_withdraws.append(event)
                 else:
                     raise AssertionError(f'Unexpected AdEx event type: {type(event)}.')
 
@@ -566,6 +676,7 @@ class Adex(EthereumModule):
             bonds=bonds,
             unbonds=unbonds,
             unbond_requests=unbond_requests,
+            channel_withdraws=channel_withdraws,
         )
 
     @staticmethod
@@ -590,10 +701,65 @@ class Adex(EthereumModule):
         """
         return {self._get_user_identity(address): address for address in addresses}
 
+    def _get_tom_pool_unclaimed_rewards(
+            self,
+            addresses: List[ChecksumEthAddress],
+    ) -> Dict[ChecksumAddress, UnclaimedReward]:
+        """Get Tom pool unclaimed rewards per address
+
+        NB: the unclaimed rewards are ADX and DAI balances associated with the
+        user identity of the address.
+        """
+        fee_rewards = self._get_tom_pool_fee_rewards_from_api()
+        identity_address_map = self._get_identity_address_map(addresses)
+        address_adx_amount = defaultdict(list)
+        address_dai_amount = defaultdict(list)
+        for entry in fee_rewards:
+            try:
+                token_ethereum_address = entry['channelArgs']['tokenAddr']
+                balances = entry['balances']
+                if token_ethereum_address == A_ADX.ethereum_address:
+                    for user_identity in identity_address_map.keys():
+                        address = identity_address_map[user_identity]
+                        address_adx_amount[address].append(FVal(balances.get(user_identity, ZERO)))
+                elif token_ethereum_address == A_DAI.ethereum_address:
+                    for user_identity in identity_address_map.keys():
+                        address = identity_address_map[user_identity]
+                        address_dai_amount[address].append(FVal(balances.get(user_identity, ZERO)))
+                else:
+                    log.error(
+                        'Unexpected token address processing AdEx unclaimed rewards. ',
+                        token_address=token_ethereum_address,
+                    )
+            except (KeyError, ValueError) as e:
+                msg = str(e)
+                log.error(
+                    'Error processing AdEx Tom pool unclaimed rewards. ',
+                    entry=entry,
+                    error=msg,
+                )
+                self.msg_aggregator.add_error(
+                    'Failed to deserialize the AdEx Tom pool unclaimed rewards. '
+                    'Check logs for details. Ignoring it.',
+                )
+                return {}
+
+        unclaimed_rewards = {}
+        for address in addresses:
+            total_adx_amount = FVal(sum(address_adx_amount[address]))
+            total_dai_amount = FVal(sum(address_dai_amount[address]))
+            # Discard addresses without amounts
+            if total_adx_amount + total_dai_amount > ZERO:
+                unclaimed_rewards[address] = UnclaimedReward(
+                    adx_amount=total_adx_amount / ADX_AMOUNT_MANTISSA,
+                    dai_amount=total_dai_amount / DAI_AMOUNT_MANTISSA,
+                )
+        return unclaimed_rewards
+
     def _get_tom_pool_incentive(self) -> Optional[TomPoolIncentive]:
         """Get Tom pool incentive data (staking rewards).
 
-        NB: the APR is the AdEx codebase and website stats APY.
+        NB: the APR calculated here is the APY in the AdEx codebase and website stats.
         https://github.com/AdExNetwork/adex-staking/blob/master/src/actions/actions.js#L86
 
         TODO: once AdEx rolls out issue #94, check how they handle incentives
@@ -601,10 +767,10 @@ class Adex(EthereumModule):
         https://github.com/AdExNetwork/adex-staking/issues/94
         """
         fee_rewards = self._get_tom_pool_fee_rewards_from_api()
-        total_staked_amount = FVal('0')
-        total_reward_per_second = FVal('0')
+        total_staked_amount = FVal(ZERO)
+        total_reward_per_second = FVal(ZERO)
         period_ends_at = Timestamp(0)
-        apr = FVal('0')
+        apr = FVal(ZERO)
         for entry in fee_rewards:
             try:
                 if entry['channelArgs']['tokenAddr'] == A_ADX.ethereum_address:
@@ -617,7 +783,7 @@ class Adex(EthereumModule):
             except (KeyError, ValueError) as e:
                 msg = str(e)
                 log.error(
-                    'Error processing AdEx Tom pool incentives.',
+                    'Error processing AdEx Tom pool incentives. ',
                     entry=entry,
                     error=msg,
                 )
@@ -628,11 +794,11 @@ class Adex(EthereumModule):
                 return None
 
         now = Timestamp(int(datetime.utcnow().timestamp()))
-        # Calculate Tom pool APR (APY)
+        # Calculate Tom pool APR
         if now < period_ends_at:  # else apr is zero
             seconds_left = FVal(period_ends_at - now)
             to_distribute = total_reward_per_second * seconds_left
-            apr = to_distribute * SECONDS_PER_YEAR / seconds_left / total_staked_amount
+            apr = to_distribute * YEAR_IN_SECONDS / seconds_left / total_staked_amount
 
         return TomPoolIncentive(
             total_staked_amount=total_staked_amount,
@@ -646,19 +812,26 @@ class Adex(EthereumModule):
             addresses: List[ChecksumEthAddress],
             from_timestamp: Timestamp,
             to_timestamp: Timestamp,
-    ) -> List[Union[Bond, Unbond, UnbondRequest]]:
+    ) -> List[Union[Bond, Unbond, UnbondRequest, ChannelWithdraw]]:
         """Returns events of the addresses within the time range and inserts/updates
         the used query range of the addresses as well.
         """
-        all_events: List[Union[Bond, Unbond, UnbondRequest]] = []
+        all_events: List[Union[Bond, Unbond, UnbondRequest, ChannelWithdraw]] = []
         for event_type_ in EventType:
-            event_type: Literal[EventType.BOND, EventType.UNBOND, EventType.UNBOND_REQUEST]
+            event_type: Literal[
+                EventType.BOND,
+                EventType.UNBOND,
+                EventType.UNBOND_REQUEST,
+                EventType.CHANNEL_WITHDRAW,
+            ]
             if event_type_ == EventType.BOND:
                 event_type = EventType.BOND
             elif event_type_ == EventType.UNBOND:
                 event_type = EventType.UNBOND
             elif event_type_ == EventType.UNBOND_REQUEST:
                 event_type = EventType.UNBOND_REQUEST
+            elif event_type_ == EventType.CHANNEL_WITHDRAW:
+                event_type = EventType.CHANNEL_WITHDRAW
             else:
                 raise AssertionError(f'Unexpected AdEx event type: {event_type_}.')
 
@@ -692,6 +865,74 @@ class Adex(EthereumModule):
             init_code=HexStr(IDENTITY_PROXY_INIT_CODE.format(signer_address=address)),
         )
 
+    def _update_staking_events_data(
+            self,
+            events: List[Union[Bond, Unbond, UnbondRequest, ChannelWithdraw]],
+    ) -> None:
+        bonds = []
+        unbonds = []
+        unbond_requests = []
+        channel_withdraws = []
+        for event in events:
+            if isinstance(event, Bond):
+                bonds.append(event)
+            elif isinstance(event, Unbond):
+                unbonds.append(event)
+            elif isinstance(event, UnbondRequest):
+                unbond_requests.append(event)
+            elif isinstance(event, ChannelWithdraw):
+                channel_withdraws.append(event)
+            else:
+                raise AssertionError(f'Unexpected AdEx event type: {type(event)}.')
+
+        # Update amounts for unbonds and unbond requests
+        bond_id_bond_map: Dict[HexStr, Optional[Bond]] = {bond.bond_id: bond for bond in bonds}
+        for event in unbonds + unbond_requests:  # type: ignore # mypy bug concatenating lists
+            has_bond = True
+            bond = bond_id_bond_map.get(event.bond_id, None)
+            if bond:
+                event.value = Balance(amount=bond.value.amount)
+                event.pool_id = bond.pool_id
+            elif event.bond_id not in bond_id_bond_map:
+                bond_id_bond_map[event.bond_id] = None
+                db_bonds = cast(List[Bond], self.database.get_adex_events(
+                    bond_id=event.bond_id,
+                    event_type=EventType.BOND,
+                ))
+                if db_bonds:
+                    db_bond = db_bonds[0]
+                    bond_id_bond_map[event.bond_id] = db_bond
+                    event.value = Balance(amount=db_bond.value.amount)
+                    event.pool_id = db_bond.pool_id
+                else:
+                    has_bond = False
+            else:
+                has_bond = False
+
+            if has_bond is False:
+                log.error(
+                    'Unexpected case updating AdEx event data. '
+                    'Unable to find the related bond event.',
+                    event=event,
+                )
+                self.msg_aggregator.add_error(
+                    'Failed to update AdEx event data. Check logs for details. Ignoring it.',
+                )
+
+        # Update usd_value for all events
+        adx_usd_price_history: Dict[Timestamp, Price] = {}
+        for event in bonds + unbonds + unbond_requests + channel_withdraws:  # type: ignore
+            adx_usd_price = adx_usd_price_history.get(event.timestamp, None)
+            if adx_usd_price is None:
+                adx_usd_price = PriceHistorian().query_historical_price(
+                    from_asset=A_ADX,
+                    to_asset=A_USD,
+                    timestamp=event.timestamp,
+                )
+
+            event.value.usd_value = adx_usd_price * event.value.amount
+            adx_usd_price_history[event.timestamp] = adx_usd_price
+
     def get_balances(
             self,
             addresses: List[ChecksumAddress],
@@ -717,15 +958,24 @@ class Adex(EthereumModule):
                     addresses=addresses,
                     event_type=EventType.UNBOND_REQUEST,
                 )
+                channel_withdraws = self._get_staking_events_graph(
+                    addresses=addresses,
+                    event_type=EventType.CHANNEL_WITHDRAW,
+                )
             except RemoteError:
                 return staking_balances
 
-            adx_usd_price = Inquirer().find_usd_price(EthereumToken('ADX'))
+            unclaimed_rewards = self._get_tom_pool_unclaimed_rewards(addresses=addresses)
+            adx_usd_price = Inquirer().find_usd_price(A_ADX)
+            dai_usd_price = Inquirer().find_usd_price(A_DAI)
             staking_balances = self._calculate_staking_balances(
                 bonds=bonds,
                 unbonds=unbonds,
                 unbond_requests=unbond_requests,
+                channel_withdraws=channel_withdraws,
+                unclaimed_rewards=unclaimed_rewards,
                 adx_usd_price=adx_usd_price,
+                dai_usd_price=dai_usd_price,
             )
         else:
             raise NotImplementedError(
@@ -745,21 +995,25 @@ class Adex(EthereumModule):
         if self.graph is None:  # could not initialize graph
             return {}
 
-        with self.trades_lock:
-            if reset_db_data is True:
-                self.database.delete_adex_events_data()
+        if reset_db_data is True:
+            self.database.delete_adex_events_data()
 
         staking_events = self._get_staking_events(
             addresses=addresses,
             from_timestamp=from_timestamp,
             to_timestamp=to_timestamp,
         )
-        adx_usd_price = Inquirer().find_usd_price(EthereumToken('ADX'))
+        unclaimed_rewards = self._get_tom_pool_unclaimed_rewards(addresses=addresses)
+        adx_usd_price = Inquirer().find_usd_price(A_ADX)
+        dai_usd_price = Inquirer().find_usd_price(A_DAI)
         staking_balances = self._calculate_staking_balances(
             bonds=staking_events.bonds,
             unbonds=staking_events.unbonds,
             unbond_requests=staking_events.unbond_requests,
+            channel_withdraws=staking_events.channel_withdraws,
+            unclaimed_rewards=unclaimed_rewards,
             adx_usd_price=adx_usd_price,
+            dai_usd_price=dai_usd_price,
         )
         tom_pool_incentive = self._get_tom_pool_incentive()
         staking_history = self._get_staking_history(