+++ rotkehlchen/chain/manager.py
@@ -919,11 +919,12 @@ class ChainManager(CacheableObject, LockableQueryObject):
                     self.totals += entry
 
         adex_module = self.adex
-        if adex_module is not None:
+        if adex_module is not None and self.premium is not None:
             adex_balances = adex_module.get_balances(addresses=self.accounts.eth)
             for address, pool_balances in adex_balances.items():
                 for pool_balance in pool_balances:
-                    eth_balances[address].assets[A_ADX] += pool_balance.balance
+                    eth_balances[address].assets[A_ADX] += pool_balance.adx_balance
+                    eth_balances[address].assets[A_DAI] += pool_balance.dai_balance
 
         # Count ETH staked in Eth2 beacon chain
         self.account_for_staked_eth2_balances(addresses=self.accounts.eth, at_addition=False)